AskdjangoBasic/requirements.txt
Package              Version
-------------------- -------
asgiref              3.4.1
Django               3.2.5
django-debug-toolbar 3.2.1
django-extensions    3.1.3
Pillow               8.3.1
pip                  21.1.3
pytz                 2021.1
setuptools           41.2.0
sqlparse             0.4.1
